+++ src/transactions/CreateWithdrawalRequestOpFrame.cpp
@@ -51,8 +51,8 @@ CreateWithdrawalRequestOpFrame::tryGetSignerRequirements(StorageHelper& storageH
 
     SignerRuleResource resource(LedgerEntryType::REVIEWABLE_REQUEST);
     resource.reviewableRequest().details.requestType(ReviewableRequestType::CREATE_WITHDRAW);
-    resource.reviewableRequest().details.withdraw().assetCode = asset->getCode();
-    resource.reviewableRequest().details.withdraw().assetType = asset->getType();
+    resource.reviewableRequest().details.createWithdraw().assetCode = asset->getCode();
+    resource.reviewableRequest().details.createWithdraw().assetType = asset->getType();
     resource.reviewableRequest().tasksToRemove = 0;
     resource.reviewableRequest().tasksToAdd = 0;
     resource.reviewableRequest().allTasks = 0;